chore: move sanity checks to common module
Moves the sanity checks to the `MdCommonModule`.

Fixes #4979.
